+++ CHANGELOG.rst
@@ -24,6 +24,12 @@ Deprecated
 
 Fixed
 -----
+- ``DirectionColorKeyTSL.direction2color()`` now accepts ``Miller`` crystal vectors as
+  inputs (in addition to ``Vector3d``).
+- ``IPFColorKeyTSL`` now correctly only accepts a single ``Vector3d`` as the sample
+  direction.
+- ``CrystalMap.shape`` no longer recomputes unless the shape has changed since the last
+  check, avoiding unnecessary recomputes that slowed down some interaction.
 
 
 2026-02-24 - version 0.14.2